requirements.txt
mysql-connector==2.2.9
mysql-connector-python==8.0.23
mysql-connector-python-rf==2.2.2
mysql-connector-repackaged==0.3.1
protobuf==3.15.8
six==1.15.0